Add number of instances inside the headings read for tests in the report  (#127)
Addresses bug 1428760.
Improves assessment report navigation by headers when using a screen reader by including the count/status in the text that's read for the header.
Does this by moving the OutcomeChip into H2 and H3. For H4, since there's additional text between the rule name and the OutcomeChip, we add the count/status to screen-reader-only text in the H4.
Replaces PR #107 and PR #121 to get past stuck license/cla check.