Fix tslint warnings in action-payloads, messages, global-action-creator (#162)
I'm touching these files in feature work and want to clean up tslint warnings, but don't want to do that in the feature PR since the changes (mostly removing the leading I from interfaces) touch a pile of files which would be a distraction for reviewers. Changes in this PR are only mechanical F2 renames.