Announce assessment in progress dialog better in NVDA (#311)
This change is a workaround for an apparent NVDA issue which would cause our "assessment in progress" dialog not to be read out as a dialog; NVDA would just read "button continue previous". This issue with alertdialogs doesn't seem specific to us or to office fabric; you can see the same behavior in W3's example alert dialog.

This change updates our implementation to, effectively, convince office fabric to set role=dialog instead of role=alertdialog, which JAWS reads no worse thasn before and NVDA reads significantly better (it reads "Assessment in progress button continue previous").

This is still not great in NVDA, but is a definite improvement and sticks within the bounds of aria recommendations. With this fix, the main issue with NVDA stops being "it never even tries to read the dialog as a dialog" and instead becomes "NVDA tries to read the full dialog contents/role as expected, but interrupts itself immediately upon the continue previous button receiving focus". Leaving any further improvement attempts for a separate PR (I didn't come up with anything that worked reliably when I experimented for many hours last night).