Improve default logger debuggability (#380)
Previously, the devtools console's file/line printout/link next to any logger-generated error or log line would always point to default-logger.ts, which is not useful. With this change, that link will instead go to the line that actually logged the message/error, which is what you actually want in 99.9% of cases.