refactor: separate e2e result publishing with separate artifacts per job (#921)
Noticed that the new e2e tests were accidentally using the same names for test run titles and failure screenshot artifacts, which made it hard to tell which was which in the pipelines results page. This refactors the common result publishing logic to a common place and makes the names/artifacts be based on job name so they'll be distinct between environments.