refactor: add missing react imports (#939)
Looking at the Typescript 3.5 update and webpack config changes it may need, I came across an compilation error where we are using React.<Mouse events> types without actually importing react on the file. I don't know why this is not happening right now with our current setup but using the experimental watch flag on one of the webpack plugins make this error appears.
Also, for sake of correctness, we should import anything we are using on each file.