test: fix Windows-hosted-agent-specific "Page crashed!" puppeteer errors (#958)
We were seeing our Windows hosted Pipelines agents hit "Page crashed!" errors somewhat reliably (~50% of e2e runs). I added functionality to collect detailed chrome logs and crash dumps and collect them as part of our published e2e results, but adding this actually fixed the issue. After some experimentation, I think the key to fixing it was in specifying a custom userDataDir launch parameter to chromium that uses a temporary userDataDir under our test-results folder, rather than deferring to Puppeteer's default behavior of putting it in the system %TEMP% directory. I'm not sure why this fixes the problem; maybe some sort of permissions issue, maybe some sort of overaggressive cleanup of the temp folder while the test is still running.

Validated that 15x runs on Windows with the userDataDir in place never saw the "Page crashed!" issue. Prior to this fix, we see it in about 50% of runs.

Out of scope: other buckets of e2e flakiness issues. In particular, it's expected that there will be some e2e-reliability failures still in the checks section of this PR (but none of them should be "Page crashed!" errors)