test: improve Windows agent yarn install time by adding network-timeout to yarnrc (#985)
The build logs from our windows agents' "install packages and dependencies" step look like this:

```
##[command]"C:\windows\system32\cmd.exe" /D /E:ON /V:OFF /S /C "CALL "d:\a\_temp\70d37892-7d53-4f2e-abc9-d9850f706082.cmd""
yarn install v1.17.3
[1/5] Validating package.json...
[2/5] Resolving packages...
[3/5] Fetching packages...
info There appears to be trouble with your network connection. Retrying...
info fsevents@1.2.9: The platform "win32" is incompatible with this module.
info "fsevents@1.2.9" is an optional dependency and failed compatibility check. Excluding it from installation.
[4/5] Linking dependencies...
[5/5] Building fresh packages...
Done in 145.07s.
##[section]Finishing: install packages and dependencies
```

Note the `info There appears to be trouble with your network connection. Retrying...` line; this doesn't appear in the mac or linux agents. The issue is that yarn uses a platform-specified network timeout, and the one on Windows is lower than mac/linux; I think what's happening here is it's hitting the timeout during the first attempt when trying to download all packages having successfully downloaded everything but the big office fabric icons package, then retrying just that one package, and having enough time for it when downloading just that one package in the retry.

Hopefully this should improve the total runtime of this step on the windows agents (it currently takes ~2x the time of the other OSes)

I sorted the .prettierignore file here; the only material change is adding the new .yarnrc file.