test: waitForSelector before evaling against it in getOuterHTMLOfSelector (#992)
e2e-reliability CI build for https://dev.azure.com/ms/accessibility-insights-web/_build/results?buildId=27762 hit a failure where it tried to getOuterHTMLOfSelector against an element that wasn't yet present on the page. This PR makes that helper more reliable by having it wait for the element to exist on the page (with the normal timeout) prior to grabbing its snapshot.