fix: avoid popup being invisible in Firefox (#1005)
For some time, the extension's popup window has used CSS that sets `body, html { height: 0px; }` as a workaround for a Chromium issue where, without that CSS, the popup window does not resize like it should when its content is zoomed in/out. It doesn't actually set the window height to 0; it forces it to dynamically fit the content.

(note that, due to a *separate* Chromium issue, you can't zoom in/out on the popup content directly; to observe this, you need to have the popup open while you're zooming in/out on a different extension page, like the details page)

In Firefox, the original workaround has the effect you would *actually* expect, and causes the popup window to be of 0 height, ie, invisible. Using `fit-content` instead of `0px` still achieves the desired dynamic zoom behavior in Chrome, but also achieves a correct popup window size in Firefox.

Note that this fix does *not* allow for correct popup window zooming in Firefox; it only enables correct window sizing in Firefox at the 100% zoom level. Popup window sizing at other zoom levels is broken in Firefox due to [Firefox bug 1570849](https://bugzilla.mozilla.org/show_bug.cgi?id=1570849), which appears to affect all extensions that use this type of popup window.