chore: replace uglifyjs with terser for minifying prod builds (#1061)
Replaces the minifier we're using on prod builds from `uglify-js` to `terser`.

Terser is a more updated replacement for UglifyJS. Webpack 4 switched to using it as the default minifier in Nov 2018 (https://github.com/webpack/webpack/pull/8392). The reasons we'd want to switch:

* Unblocks future updates to typescript output target; uglify only supports up to ES5 javascript, Terser supports more recent javascript as input
* Closer to the "normal" webpack happy path
* As of recently, [more widely used (with webpack)](https://www.npmtrends.com/uglifyjs-webpack-plugin-vs-terser-webpack-plugin)

It is essentially a drop in replacement; we expect no customer-visible changes as a result of this update.