refactor: replace NodeSelector with Document in preparation for Typescript upgrade (#1065)
Updating from Typescript 2.9 to 3.5 results in a variety of errors that look like this:

```
ERROR in Q:/repos/accessibility-insights-web/src/common/document-manipulator.ts(4,42):
TS2304: Cannot find name 'NodeSelector'.
```

It looks like https://github.com/microsoft/TSJS-lib-generator/pull/647 was the point where it was removed, merged into a separate `ParentNode` interface. We considered using `ParentNode` instead here, but `Document` seemed closer to the intended usage in most cases and led to the tests working more similarly to the product code than anything else, so that's what I went with.

Replacing the usages with Document is a no-op in Typescript 2.9 (so nice to do as a separate PR, to make the eventual Typescript upgrade PR smaller). It also forces the affected tests to be a little bit better (they create test documents for the cases where the product passes real documents, rather than creating test nodes within a global document and hoping that they act similarly and that no other test touches global document state).