fix: prevent redundant assessment overview page scroll bars (#1098)
Second half of fixing #1088. A regression between chrome 75 and 76 broke the scrolling behavior of the assessment overview page, resulting in multiple redundant scrollbars and a malformed size on the left nav area. This addresses both issues.

This does *not* address the test details rows being misaligned; see separate PR #1096 for that.

Before:
![image](https://user-images.githubusercontent.com/376284/63310043-feb64680-c2ad-11e9-835d-a22d8c71206c.png)

After:
![image](https://user-images.githubusercontent.com/376284/63310081-29080400-c2ae-11e9-9a63-e02a0905bbdf.png)