refactor: IssueDetailsTextGenerator to use IssueDetailsBuilder (#1117)
`IssueDetailsTextGenerator` is used on the "copy failure details" button but it has duplicated logic (that we have in a generic implementation under issue filing for git hub and azure boards).
This PR refactors `IssueDetailsTextGenerator` to use `IssueDetailsBuilder` instead of having it's own logic about the issue details content. In order to do that, I had to update `IssueDetailsBuilder` to account for some style differences between file a bug and copy failure details. I've also add a new markup formatter: plain text formatter.
After this refactor, the main section of the issue details will be hold on one place only: `IssueDetailsBuilder`

* use destructuring on MarkupFormatter param
* add 2 markup formatter functions
* use new markup functions on issue details builder
* add newLine function to MarkupFormatter
* use newLine on issue details builder
* add plain text formatter
* use IssueDetailsBuilder for the IssueDetailsTextGenerator
* update plain text formatter for link
* rename newLine to sectionSeparator