Revert device connect state if port number changes (#1123)
Currently in the device connect screen, if a user enters a port, validates it, and then changes the port, the state remains unchanged. That behavior makes it seem like the user has validated their newly entered port, which is not the case. This PR updates the behavior such that whenever the user changes the port value, the state is restored to the original, pre-validation state. There are no actual UI/visual changes.