refactor: minor clean up on components/tests (#1222)
While looking at the electron app code to get familiar before starting a new feature, @JGibson2019 and I did some small clean up refactoring on components and tests.

One thing that caught our attention: `DeviceConnectPortEntry` uses an `async` `onClick` function for the primary `Button` it uses. I think this is the first time I see and async onClick function on our code base. Looking at the tests, we found a call to `setInmediate`. Again, this was odd as is the first time I see this on our code base. Playing around with the test code, we figure out this two things were related (async onClick function and setInmediate on the test). Because the onClick function is async, you need to wait (somehow) for it to finish before asserting on the tests (e.g calling `verifyAll` on a mock) but enzyme doesn't expose onClick handlers in a way that we can directly use `await` or `then`, hence, the use of `setInmediate`. We decided to change this a little bit (after some discussion with @dbjorge) to match our code style better. We introduce a `tick` function to wrap the waiting code (the call to `setInmediate` although, we end up using `setTimeout` with 0 ms delay). This way we can write `await tick();` and not have to either inline a function inside `setInmediate` (which kind of looks odd because of indentation, etc) or create a private function of sort to pass to `setInmediate`.

* small clean up on test
* Make onChange function a private function (instead of a const inside render function)
* general clean up. improve render test
* improve port text field test
* improve validate port number test
* reduce duplication
* Using .each	able instead of passing an array

Co-authored-by: JGibson2019