refactor: electron external link props (#1225)
Updated `ElectronExternalLinkProps` to contain 'children' property instead of 'text' so its usage is similar to that of `NewTabLink`. This is one of the steps to allow for the reuse of the `TelementryPermissionDialog`, which contains the `telemetryPopupNotice` that uses `NewTabLink`. Will update `telemetryPopupNotice` to be able to use `NewTabLink` or `ElectronExternalLink` in a future PR.

* replaced ElectronExternalLinkProps 'text' with 'children'
* update snapshots
* update snapshot
* fix props.children usage