Make PayloadCallback a more strict type (#1248)
#### Description of changes

_**Note**: This is part of the Telemetry Opt-In feature for electron AI_

Make `PayloadCallback` type use a template type for the `payload` parameter. This way we have type info when calling `RegisterTypeToPayloadCallback` (usually on the action creators).

This is part of a bigger refactor where we're trying to make the action message creator and action creator to have the same public API, this way we can use one or the other on the view. The use case for this is the telemetry opt in dialog, which is currently being use on the popup and we want to reuse on the electron app.

#### Pull request checklist

- [x] Addresses an existing issue: part of WI # 1597954
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [x] (UI changes only) Added screenshots/GIFs to description above
   - not affected
- [x] (UI changes only) Verified usability with NVDA/JAWS
   - not affected