refactor: wrap ElectronExternalLink to not need shell prop (#1262)
While working on #1258 I found out we do not have a suitable candidate for the `LinkComponent` dependency of the telemetry opt in. The plan was to reuse `ElectronExternalLink` but this component expect a `shell` prop (which comes from the electron api).

This PR adds a wrapper component around the `ElectronExternalLink` so we don't need to pass the `shell` prop and we can use it for the `LinkComponent` dependency.