Add automated checks command bar (#1345)
The ongoing automated checks results from Android feature calls for the addition of a command bar. For now, the only command with an actual purpose is the rescan button. We could add the rest now, but they will not do anything until future features are complete. We don't yet have the UI to house this bar, so for now it is just waiting to be used.