refactor: remove unnecessary details view div/styles (#1472)
#### Description of changes

The details view container currently adds a `<div className="table column-layout main-wrapper">` div near the root level. `main-wrapper` is never used in our css, and `table column-layout` is better represented as applying styles to a more logically-named class (so the CSS can have layout responsibility, rather than essentially using inline styles here).

This removes the unnecessary layer of div as well as the `.table` classes from `detailsview.scss` (this was the last spot they were used - there is a separate `.table` style in `content.scss` that is a little more reasonble, which I haven't modified in this PR.

Verified that there are no visual changes as a result of this change. In particular, verified the following details view scenarios:

* FastPass (both left nav tabs)
* Assessement (overview, manual, assisted requirements)
* Settings panel
* Issues filing dialog
* "No content available" page

#### Pull request checklist

- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS