Add code to propigate the CardSelectionStoreData to the card view model
This change adds the data from the CardSelectionStore into the cards view model. The store data is first transformed into data relevant to the view model by a call to getCardSelectionViewData. The CardSelectionViewData is then passed to getUnifiedRuleResults where it is assigned to newly added fields in the view model data. The data can now be consumed by the details view.