feat(android-cards-view): wire up CardsView (#1496)
Wiring up `CardsView` on ai-android.

**Notes**
I'm purposfully creating the `CardsViewDeps` as a separated object so we know which dependencies are needed from `CardsView` (or child components). I left some comments, on some properties, specially on issue filing related deps. I want the clean up part of the deps to be on a separated PR so we can more thoroughly check what we actually need and what we don't

**Out of Scope**
I only added `CardRow` components for the most simple values, basically of type string. That is: `text`, `className` and `componentDescription`. There is another PR to take care of `howToFix` (#1489)