feat(screenshot): create screenshot highlight box component (#1596)
Created a highlight box component that will be used to visually highlight elements on the screenshot. The highlight boxes will render on top of the screenshot image.

Important note: The highlight box will position itself relative to its parent component due to the position: absolute property, but only when its parent component has position: relative. This is necessary to simplify rendering the highlight boxes on top of a screenshot image. The screenshot container that will render the screenshot image and all highlight boxes on top will need to include position: relative in its properties.