refactor: promisify inject api calls (#1640)
As part of the permissions works, I'd like to convert `executeScriptInTab` and `insertCSSInTab` from using callbacks to use promises. This will help on how we handle permissions when injecting in a future PR.

As a note: we already converted `getUserData`, `setUserData` and `removeUserData` (all part of `StorageAdpater` type) successfully in the past.