feat(consistent-cards-UI): Update card selection view data with visual helper (#1646)
The intent of the change is that when the visual helper is disabled, no cards are selected and no instances are highlighted. Rule expansion should remain unchanged.

Note: the change to rule-based-view-model-provider.test.ts was only to add the visualHelperEnabled field because the code in rule-based-view-model-provider.ts only returns true for this field... for now... watch this space.