feat(and-screenshot): make screenshot+highlights rescale dynamically to fit their pane (#1641)
#### Description of changes

Before this PR, the screenshot and highlights were rendered at actual size, using both horizontal and vertical scrollbars if that was bigger than the pane containing them, like this:

![screenshot of screenshot not resizing with scrollbars](https://user-images.githubusercontent.com/376284/68423937-d83ed380-0157-11ea-8494-67e05abdc559.png)

After this PR, the screenshot and highlights both resize dynamically to fit the pane containing them, like this:

![screenshot of screenshot resizing to fit pane](https://user-images.githubusercontent.com/376284/68424072-176d2480-0158-11ea-9e19-31ebf1aac9b1.png)

Making the screenshot size itself dynamically was straightforward, just a bit of CSS (in particular, `object-fit: contain` to deal with maintaining the aspect ratio).

Making the highlight boxes respect the size was trickier. I considered using a CSS scale transform, but we didn't want the box outline width or the `!` sizes to change, just their positions, so that wasn't going to work. To achieve the desired effect, I switched the boxes from calculating positions as absolute px offsets relative to their container div to instead calculate relative % offsets relative to the container div. Since this required more a more interesting calculation involving knowledge of the viewport size, I decided to move the calculation out of the HighlightBox component and into the ViewModel, and to calculate it in the ViewModelProvider; updating the piping from ViewModelProvider -> HighlightBox accordingly is where most of the code change in the PR comes from.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: 1640687 (partial)
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). Check workflow guide at: `<rootDir>/docs/workflow.md`
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS