refactor: Add all possible scenarios to rule based view model provider unit tests (#1647)
If you look at rule-based-view-model-provider.test.ts before this change, you will see that only three of a possible 8 test scenarios were being tested.

Admittedly, some test scenarios were unexpected, such as
{ isExpanded: false, isSelected: true, isHighlighted: false },
because this possibility __should__ never be produced by getCardSelectionViewData.

However, valid test scenarios such as
{ isExpanded: true, isSelected: false, isHighlighted: true },
and
{ isExpanded: false, isSelected: false, isHighlighted: false }
were never tested.

I Assert that unit tests are not just for testing the things we expect, but also for testing the things we don' t expect. Therefore, I have added code to create all the possible input scenarios.

Furthermore, I have added the 8 possible scenarios for when the newly added CardSelectionViewData.visualHelperEnabled field is present, bringing the total number of scenarios up to 16.

The astute observer will note that, for the moment, the visualHelperEnabled field of cards view data is always set to true. I'm not touching the production code which sets visualHelperEnabled in this PR so that in the next PR, the changes in the unit test snapshots will be clearly visible as a result of changing the production code.

The last thing to note about this change is that while the unit tests have grown, the production code has not changed; and we have validated the production code empirically. Therefore, the new snapshots added in this change are, practically speaking, representative of the expected behavior.