refactor: introduce injection action creator (#1670)
As part of the browser extension permissions effort, we need a to make some changes on how and when we inject our scripts into the target page.

In this refactor, the injection related code is move out of the `ActionCreator` into a more specific action creator. A new `InjectionActions` class has been created and all related classes (`VisualizationStore` is notable here) have been updated.