test: add background page window.insightsUserConfiguration controller for E2Es to use (#1728)
#### Description of changes

This PR introduces a new controller on the background page console similar to the existing `window.insightsFeatureFlags` controller named `window.insightsUserConfiguration`. We want this primarily so that e2e tests can be updated to use it rather than UI to do most user configuration tasks (particularly, dismissing the initial telemetry prompt and setting high contrast mode), since the current strategy of setting high contrast mode via the UI has the following repercussions:

* causes issues for the upcoming permission changes (many tests attempt to toggle details view settings from states that will soon be "tab closed" pages)
* takes a lot of unnecessary time
* causes a lot of unnecessary instability

This PR introduces the new controller, updates existing browser setup to use it for suppressing the telemetry dialog, and adds a new `settings-panel.test.ts` to verify that the new config controller works in sync with the settings panel UI (so we can still keep the test coverage verifying that the high contrast toggle actually works).

Despite adding a new file of e2e tests, this PR cuts off about a minute (~12%) off the e2e runtime on build agents. On my dev box, the followup change to use this strategy for high contrast settings cuts off another few minutes (~35%).

Out of scope: updating all existing e2e tests to use the new infrastructure for high contrast mode toggling. That turns out to be a large change on its own since it updates a bunch of tests to stop doing so much redundant "normal" vs "high-contrast" setup and changes a lot of snapshots accordingly, so l extracted it for a later PR.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). Check workflow guide at: `<rootDir>/docs/workflow.md`
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS