refactor: update all lodash imports (#1777)
Update all usages of `import * as _ from 'lodash'` to explicitly import what we are actually using from lodash. This should impact the bundle size in a positive way as the former style ends up including all of `lodash` code, while the latter helps webpack with the tree shaking. I do not have numbers for this though.

* Change all lodash imports to not use *
* alphabetize imports
* use find from lodash not office fabric