fix: use class name from scss.d file (#1787)
We are hardcoding a class name for `RulesOnly` component. This works fine on "dev" mode builds but breaks on "production" mode builds. The main reason is: on production mode, we add a hash as a suffix to all class names (to avoid name collision) while on dev mode we don't change the class names at all.

Using the class name from the corresponding scss.d.ts file fix this issue.

* fix class name by importing it from scss file
* update snapshot