refactor: remove allTests from DetailsViewPivotType enum (#1798)
`allTest` was associated with a view (on the details view) where we use to show info about each of the visualization we actually have on the ad hoc tools (so, automated checks, headings, landmarks, color and tab stops) but it got "remove" in favor of having just fast pass and assessment. Yet, the enum value persist on our code along side some small pieces of code to handle the value. Since we don't show that view on the details view anymore, it does not make sense to even have the enum value neither.

In this PR I removed all references to `allTest`, update a couple of pieces of logic for handling that value and update all the test to not reference it (I end up using either `fastPass` or `assessments` depending on the context)

* remove logic that uses allTests on visualization store
* use assessment instead of allTest
* remove the reset of the references to allTest